+++ src/main/java/pl/pjatk/MovieService/service/MovieService.java
@@ -12,18 +12,18 @@ import static pl.pjatk.MovieService.model.Category.Horror;
 public class MovieService {
 
     public List<Movie> findAll() {
-        return List.of(new Movie(1, "Straszny Film", Horror, "somebody", 5));
+        return List.of(new Movie(1L, "Straszny Film", Horror, "somebody", 5));
     }
 
     public Movie findById() {
-        return new Movie(1, "Straszny Film", Horror, "somebody", 5);
+        return new Movie(1L, "Straszny Film", Horror, "somebody", 5);
     }
 
     public Movie addMovie(Movie movie){
         return movie;
     }
 
-    public Movie editMovie(int id, Movie movie) {
+    public Movie editMovie(Long id, Movie movie) {
         return new Movie(id, movie.getTitle(), movie.getCategory(), movie.getDirector(), movie.getRating());
     }
